feat: standalone mandate-keeperhub-adapter crate (Stretch IP-4)
Builds the publishable IP-4 surface from docs/keeperhub-integration-paths.md.
The IP-4 prereq trait/types extract landed in PR #49 (mandate-core::execution
hosts GuardedExecutor, ExecutionReceipt, ExecutionError, MandateEnvelope).
This PR moves the actual KeeperHubExecutor / KeeperHubMode / KeeperHubLiveConfig
implementation out of crates/mandate-execution::keeperhub and into a new
top-level crate, crates/mandate-keeperhub-adapter, that depends ONLY on
mandate-core (plus utility deps: serde, thiserror, ulid, hex). No leaks
into mandate-execution, mandate-server, mandate-storage, mandate-cli,
mandate-policy, or mandate-mcp — verified at function-signature level.

What ships:

- crates/mandate-keeperhub-adapter/
    Cargo.toml          — publishable metadata (license, repository,
                          description, keywords, categories, readme)
    src/lib.rs          — KeeperHubExecutor + KeeperHubMode +
                          KeeperHubLiveConfig moved here; re-exports
                          GuardedExecutor + ExecutionReceipt +
                          ExecutionError + MandateEnvelope from
                          mandate_core::execution
    examples/submit_signed_receipt.rs
                        — runnable example: APRP + receipt + envelope
                          construction + (BackendOffline) submission
    README.md           — install via cargo add, quickstart, "what this is
                          NOT" section, links to integration-paths +
                          live-spike docs
    CHANGELOG.md        — v0.1.0 initial release notes

- crates/mandate-execution: keeperhub.rs deleted; lib.rs re-exports the
  adapter's KeeperHub types so existing consumers (mandate-server,
  mandate-cli, demo-agents) compile without import-path changes.

- Workspace Cargo.toml: new member added.

- Docs refresh per the auditor's polish: IP-4 framing across
  docs/keeperhub-integration-paths.md, keeperhub-live-spike.md,
  partner-onepagers/keeperhub.md, audit doc, CLI mcp doc, product
  backlog, README, IMPLEMENTATION_STATUS, SUBMISSION docs. Numerical
  drift refreshed: 308 → 310 cargo tests (the +2 are new
  mandate-keeperhub-adapter tests).

What this is NOT:

- Not actually published on crates.io. `cargo publish` would require
  mandate-core to be published first (cascade requirement); this is a
  separate post-hackathon decision tracked in CHANGELOG / README.
  The adapter is publishable in shape (correct metadata, no internal
  leaks beyond mandate-core, examples build, README ready).
- Not a live KeeperHub client. `KeeperHubExecutor::live()` still returns
  BackendOffline until a stable submission schema and credentials wire.
  Doc framing matches: "publishable workspace crate today; crates.io
  publication target."

Verification (per Round 0 + this commit):

- cargo fmt --check: clean
- cargo clippy --workspace --all-targets -- -D warnings: clean
- cargo test --workspace --all-targets: 310 / 310 pass (308 baseline + 2
  new in mandate-keeperhub-adapter tests)
- cargo audit --deny warnings: clean
- bash demo-scripts/run-openagents-final.sh: 13/13
- bash demo-scripts/run-production-shaped-mock.sh: 26/0/1
- bash demo-scripts/sponsors/mcp-passport.sh: green
- python3 trust-badge/test_build.py: 49/49
- python3 operator-console/test_build.py: 118/118
- All existing consumers compile + test pass; back-compat re-export from
  mandate-execution preserves ALL existing import paths.
- cargo publish --dry-run -p mandate-keeperhub-adapter: blocked on
  mandate-core not yet on crates.io (expected; documented as a
  publication cascade requirement, not a structural defect).
