+++ IMPLEMENTATION_STATUS.md
@@ -57,9 +57,9 @@ Full Open Agents vertical:
 - Active-policy lifecycle: `mandate policy {validate, current, activate, diff}` (PSM-A3) backed by SQLite migration V006 (`active_policy` table with partial UNIQUE singleton index). Local lifecycle, not remote governance — `docs/cli/policy.md` documents the scope honestly.
 - Static, offline trust-badge proof viewer (`trust-badge/build.py`, stdlib Python) + stdlib regression test (`trust-badge/test_build.py`). No JS, no fetch, works from `file://`.
 - **Mandate Passport capsule schema + verifier + CLI** (`schemas/mandate.passport_capsule.v1.json`, `crates/mandate-core/src/passport.rs`, `crates/mandate-cli/src/passport.rs`, Passport P1.1 + P2.1). `mandate passport run` drives the existing `POST /v1/payment-requests` pipeline in-process, reads back the audit event, builds a checkpoint, and self-verifies the capsule before atomic write. `mandate passport verify` runs the structural verifier against any capsule. 8 tampered fixtures pin every cross-field invariant.
-- **MCP-callable Mandate gateway** (`crates/mandate-mcp/`, Passport P3.1) — stdio JSON-RPC 2.0 server exposing `mandate.validate_aprp`, `mandate.decide`, `mandate.run_guarded_execution`, `mandate.verify_capsule`, and `mandate.audit_lookup` (the IP-3 sister tool to KeeperHub's proposed `keeperhub.lookup_execution`). 22 integration tests across in-process dispatch and stdio child-process transport. Judge-facing walk-through: [`docs/mcp-integration-guide.md`](docs/mcp-integration-guide.md) (Passport P3.2). Sponsor demo: `bash demo-scripts/sponsors/mcp-passport.sh` (writes `demo-scripts/artifacts/mcp-transcript.json`).
-- **KeeperHub `mandate_*` envelope helper** (`crates/mandate-execution/src/keeperhub.rs::KeeperHubEnvelope`, Passport P5.1, IP-1) — composes `mandate_request_hash` + `mandate_policy_hash` + `mandate_receipt_signature` + `mandate_audit_event_id` (target: `mandate_passport_capsule_hash`) directly from an existing `PolicyReceipt` so a `KeeperHubExecutor::live()` body can drop it onto the workflow webhook submission. **Helper shipped; the live HTTP wiring is still gated on KeeperHub publishing a stable submission/result schema** (see `docs/keeperhub-live-spike.md` §Open questions).
-- **GuardedExecutor / ExecutionReceipt / ExecutionError moved to `mandate-core`** (`crates/mandate-core/src/execution.rs`, post-P5.1 IP-4 prerequisite, [#49](https://github.com/B2JK-Industry/mandate-ethglobal-openagents-2026/pull/49)) — re-exported from `mandate-execution::lib` for back-compat. The standalone `mandate-keeperhub-adapter` crate that an `cargo add` consumer would pull (full IP-4) is still target; the type-move that unblocks it is on `main`.
+- **MCP-callable Mandate gateway** (`crates/mandate-mcp/`, Passport P3.1) — stdio JSON-RPC 2.0 server exposing `mandate.validate_aprp`, `mandate.decide`, `mandate.run_guarded_execution`, `mandate.verify_capsule`, and `mandate.audit_lookup` (the IP-3 sister tool to KeeperHub's proposed `keeperhub.lookup_execution`). 29 integration tests across in-process dispatch, stdio child-process transport, and path-sandbox escapes. Judge-facing walk-through: [`docs/mcp-integration-guide.md`](docs/mcp-integration-guide.md) (Passport P3.2). Sponsor demo: `bash demo-scripts/sponsors/mcp-passport.sh` (writes `demo-scripts/artifacts/mcp-transcript.json`).
+- **KeeperHub `mandate_*` envelope helper** (`mandate_keeperhub_adapter::build_envelope`, Passport P5.1, IP-1) — composes `mandate_request_hash` + `mandate_policy_hash` + `mandate_receipt_signature` + `mandate_audit_event_id` (target: `mandate_passport_capsule_hash`) directly from an existing `PolicyReceipt` so a `KeeperHubExecutor::live()` body can drop it onto the workflow webhook submission. **Helper shipped; the live HTTP wiring is still gated on KeeperHub publishing a stable submission/result schema** (see `docs/keeperhub-live-spike.md` §Open questions).
+- **Standalone `mandate-keeperhub-adapter` workspace crate** (`crates/mandate-keeperhub-adapter/`, IP-4) — exposes `KeeperHubExecutor`, `KeeperHubMode`, `build_envelope`, and re-exports `GuardedExecutor` / `ExecutionReceipt` / `ExecutionError` / `MandateEnvelope` from `mandate-core`; `mandate-execution` re-exports it for back-compat. Crates.io publication remains a target, not a shipped claim.
 - **GitHub Pages public proof site** (`.github/workflows/pages.yml` + `site/index.html`, Passport P7.1) — deploys from `main`. Renders the trust-badge + operator-console + a downloadable Passport capsule into a stable static URL. Plain HTML, no JS, no client-side network calls, no external asset; byte-grep-clean against the same safe-host allowlist as `demo-fixtures/test_fixtures.py`.
 
 ## Surfaces a judge can verify
