+++ SUBMISSION_NOTES.md
@@ -36,10 +36,10 @@ All implementation code in this repository:
 - Audit checkpoints + mock anchoring: `mandate audit checkpoint {create, verify}` (PSM-A4) backed by SQLite migration V007 (`audit_checkpoints` table). **Mock anchoring**, NOT real onchain anchoring — `mock_anchor_ref` is a deterministic local id, never broadcast. Every CLI line carries a `mock-anchor:` prefix and every JSON artifact carries `mock_anchor: true`. Documented in `docs/cli/audit-checkpoint.md`.
 - **Mandate Passport capsule schema + verifier + CLI** (`schemas/mandate.passport_capsule.v1.json`, `crates/mandate-cli/src/passport.rs`, Passport P1.1 + P2.1) — `mandate passport run` drives the existing `POST /v1/payment-requests` pipeline in-process, builds a checkpoint, and self-verifies before atomic write; `mandate passport verify` runs the structural verifier against any capsule.
 - **MCP-callable Mandate gateway** (`crates/mandate-mcp/`, Passport P3.1) — stdio JSON-RPC 2.0 server with `mandate.validate_aprp` / `mandate.decide` / `mandate.run_guarded_execution` / `mandate.verify_capsule` / `mandate.audit_lookup` (the IP-3 sister tool to KeeperHub's proposed `keeperhub.lookup_execution`). Sponsor demo: `bash demo-scripts/sponsors/mcp-passport.sh`. Judge-facing walk-through: `docs/mcp-integration-guide.md` (Passport P3.2).
-- **KeeperHub `mandate_*` envelope helper** (`crates/mandate-execution/src/keeperhub.rs::KeeperHubEnvelope`, Passport P5.1, IP-1) — composes `mandate_request_hash` / `mandate_policy_hash` / `mandate_receipt_signature` / `mandate_audit_event_id` directly from an existing `PolicyReceipt`. Helper shipped; live KeeperHub HTTP wiring still gated on KeeperHub publishing a stable submission/result schema (open questions in `docs/keeperhub-live-spike.md`).
-- **GuardedExecutor / ExecutionReceipt / ExecutionError moved to `mandate-core`** ([#49](https://github.com/B2JK-Industry/mandate-ethglobal-openagents-2026/pull/49), IP-4 prerequisite) — re-exported from `mandate-execution::lib` for back-compat. The standalone `mandate-keeperhub-adapter` crate (full IP-4) is still target; the type-move that unblocks it is on `main`.
+- **KeeperHub `mandate_*` envelope helper** (`mandate_keeperhub_adapter::build_envelope`, Passport P5.1, IP-1) — composes `mandate_request_hash` / `mandate_policy_hash` / `mandate_receipt_signature` / `mandate_audit_event_id` directly from an existing `PolicyReceipt`. Helper shipped; live KeeperHub HTTP wiring still gated on KeeperHub publishing a stable submission/result schema (open questions in `docs/keeperhub-live-spike.md`).
+- **Standalone `mandate-keeperhub-adapter` workspace crate** (IP-4) — `KeeperHubExecutor`, `GuardedExecutor` re-exports, `build_envelope`, README, changelog and runnable example live under `crates/mandate-keeperhub-adapter/`; `mandate-execution` re-exports it for back-compat. Crates.io publication remains a target, not a shipped claim.
 - **GitHub Pages public proof site** (`.github/workflows/pages.yml` + `site/index.html`, Passport P7.1) — deploys from `main` to `https://b2jk-industry.github.io/mandate-ethglobal-openagents-2026/`. Plain HTML, no JS, no client-side network calls; same offline-only ground rules as trust-badge / operator-console.
-- CI: fmt, clippy, tests (300 passing), schema validation, OpenAPI validation, trust-badge regression test, operator-console regression test, demo-fixtures validator.
+- CI: fmt, clippy, tests (310 passing), schema validation, OpenAPI validation, trust-badge regression test, operator-console regression test, demo-fixtures validator.
 - Production-shaped mock runner: `bash demo-scripts/run-production-shaped-mock.sh` walks the operator surface end-to-end (doctor, mock KMS CLI, full PSM-A3 policy lifecycle, **PSM-A4 audit checkpoint create/verify with mock anchoring**, persistent SQLite allow + deny, audit-bundle export, Passport P2.1 capsule emit/verify, plus the `mandate-operator-evidence-v1` transcript consumed by the operator console). Tally: **26 real / 0 mock / 1 skipped** — every A-side backlog row has merged; only the optional `--include-final-demo` flag remains on the SKIPPED list.
 - Self-audit artefact: [`SECURITY_NOTES.md`](SECURITY_NOTES.md) documents the known scope limitations a production deployment would need to address (daemon authentication, production signer wiring, budget tracker persistence, idempotency in-flight semantics, Passport verifier scope). Honest disclosure, not a roadmap promise.
