+++ crates/mandate-execution/Cargo.toml
@@ -9,6 +9,10 @@ description = "Mandate execution: sponsor execution adapters (KeeperHub, Uniswap
 
 [dependencies]
 mandate-core = { workspace = true }
+# IP-4: the KeeperHub adapter lives in its own publishable crate. We
+# pull it in here so existing call sites that import via
+# `mandate_execution::KeeperHubExecutor` keep resolving unchanged.
+mandate-keeperhub-adapter = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
