+++ crates/mandate-keeperhub-adapter/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "mandate-keeperhub-adapter"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+repository = "https://github.com/B2JK-Industry/mandate-ethglobal-openagents-2026"
+description = "Mandate adapter for KeeperHub workflow execution: signed PolicyReceipt + IP-1 mandate_* envelope, GuardedExecutor trait, mock + live constructors. See repo for the IP-1..IP-5 catalogue."
+keywords = ["mandate", "keeperhub", "policy", "agent", "execution"]
+categories = ["api-bindings", "cryptography"]
+readme = "README.md"
+rust-version.workspace = true
+
+[dependencies]
+# Mandate core hosts the IP-4 trait (`GuardedExecutor`) + receipt + envelope
+# types. This is the *only* workspace-internal dependency, by design — the
+# whole point of this crate is "depend on mandate-keeperhub-adapter" without
+# pulling the rest of the Mandate workspace.
+mandate-core = { workspace = true }
+
+serde = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+ulid = { workspace = true }
+hex = { workspace = true }
+
+[dev-dependencies]
+chrono = { workspace = true }
