+++ crates/mandate-keeperhub-adapter/examples/submit_signed_receipt.rs
@@ -1,0 +1,78 @@
+//! Standalone example: build a signed `PolicyReceipt`, hand it to the
+//! KeeperHub executor, observe the IP-1 envelope and the
+//! `BackendOffline` live-mode response.
+//!
+//! Run with: `cargo run --example submit_signed_receipt -p mandate-keeperhub-adapter`
+//!
+//! This example imports ONLY `mandate_keeperhub_adapter` and
+//! `mandate_core` — it doesn't touch `mandate-server`, `mandate-policy`,
+//! `mandate-storage`, `mandate-cli`, or `mandate-mcp`. That's the
+//! IP-4 promise: a third-party agent framework can do the same.
+
+use mandate_core::aprp::PaymentRequest;
+use mandate_core::receipt::{
+    Decision, EmbeddedSignature, PolicyReceipt, ReceiptType, SignatureAlgorithm,
+};
+use mandate_keeperhub_adapter::{build_envelope, GuardedExecutor, KeeperHubExecutor};
+
+fn main() {
+    // 1. Build a request body. In a real agent, this is the APRP the
+    //    agent sent through Mandate's HTTP / MCP surface; here we
+    //    construct a placeholder that satisfies the schema enough to
+    //    flow through the executor.
+    let aprp_json = include_str!("../../../test-corpus/aprp/golden_001_minimal.json");
+    let request: PaymentRequest =
+        serde_json::from_str(aprp_json).expect("APRP fixture is schema-valid JSON");
+
+    // 2. Construct a signed PolicyReceipt. In production, this comes
+    //    back from `mandate_server::router` after the policy +
+    //    budget + nonce + audit pipeline has run. For an offline
+    //    example we hand-build one that the executor will accept.
+    let receipt = PolicyReceipt {
+        receipt_type: ReceiptType::PolicyReceiptV1,
+        version: 1,
+        agent_id: "research-agent-01".into(),
+        decision: Decision::Allow,
+        deny_code: None,
+        request_hash: "a".repeat(64),
+        policy_hash: "b".repeat(64),
+        policy_version: Some(1),
+        audit_event_id: "evt-01HTAWX5K3R8YV9NQB7C6P2DGS".into(),
+        execution_ref: None,
+        issued_at: chrono::Utc::now(),
+        expires_at: None,
+        signature: EmbeddedSignature {
+            algorithm: SignatureAlgorithm::Ed25519,
+            key_id: "decision-signer-v1".into(),
+            signature_hex: "f".repeat(128),
+        },
+    };
+
+    // 3. Inspect the IP-1 envelope that the live arm would carry. This
+    //    is the wire payload future KeeperHub workflow webhooks
+    //    receive alongside the APRP body and signed receipt.
+    let envelope = build_envelope(&receipt);
+    println!("== IP-1 envelope (canonical wire form) ==");
+    println!("{}", envelope.to_json_payload());
+
+    // 4. Mock execution — succeeds with a fresh kh-<ULID> ref.
+    let mock = KeeperHubExecutor::local_mock();
+    let mock_receipt = mock
+        .execute(&request, &receipt)
+        .expect("local_mock always succeeds on Decision::Allow");
+    println!("\n== mock execute result ==");
+    println!(
+        "sponsor={} mock={} execution_ref={}",
+        mock_receipt.sponsor, mock_receipt.mock, mock_receipt.execution_ref
+    );
+
+    // 5. Live execution — would call KeeperHub today. Returns
+    //    BackendOffline because credentials aren't wired. The IP-1
+    //    envelope is still constructed inside the live arm, just not
+    //    sent.
+    let live = KeeperHubExecutor::live();
+    match live.execute(&request, &receipt) {
+        Ok(_) => unreachable!("live mode currently always returns BackendOffline"),
+        Err(e) => println!("\n== live execute (gated) ==\n{e}"),
+    }
+}
